+++ changes/changelog.md
@@ -8,7 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 - Changed the build script to use volumes automatically.
+
+### Fixed
 - Fixed double-stringify when persisting JSON.
+- Fixed the user to remind
 
 ## [1.3.3] - 2019-07-22
 ### Added