+++ changes/changelog.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Fixed
 - Fixed double-stringify when persisting JSON.
 - Fixed the user to remind
+- Reminders can now have full strings
 
 ## [1.3.3] - 2019-07-22
 ### Added